app/content-creator/requirements.txt
flask==2.3.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-exporter-prometheus==0.41b0
prometheus-client==0.17.1
requests==2.31.0
